added fb links and author

+++ src/newsletter/views.py
@@ -1,15 +1,20 @@
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect,Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from .forms import PostForm
 from .models import Post
 # Create your views here.
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    if not request.user.is_authenticated:
+        raise Http404
     form = PostForm(request.POST or None,request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request,"Successfully created")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -39,7 +44,7 @@ def post_list(request): #list items
 
 def post_articles(request): #list items
     queryset_list = Post.objects.all()
-    paginator = Paginator(queryset_list, 3) # Show 25 contacts per page
+    paginator = Paginator(queryset_list, 6) # Show 25 contacts per page
     page = request.GET.get('page')
     try:
         queryset = paginator.page(page)
@@ -57,6 +62,8 @@ def post_articles(request): #list items
 
 
 def post_update(request,id=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance=get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None,instance=instance)
     if form.is_valid():
@@ -72,6 +79,8 @@ def post_update(request,id=None):
     return render(request,"post_form.html", context)
 
 def post_delete(request,id=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
     instance=get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request,"successfully deleted")
